+++ include/metashell/engine_entry.hpp
@@ -18,6 +18,7 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 #include <metashell/data/feature.hpp>
+#include <metashell/data/markdown_string.hpp>
 #include <metashell/engine_factory.hpp>
 
 #include <boost/filesystem/path.hpp>
@@ -32,7 +33,7 @@ namespace metashell
   public:
     engine_entry(engine_factory factory_,
                  std::string args_,
-                 std::string description_,
+                 data::markdown_string description_,
                  std::vector<data::feature> features_);
 
     std::unique_ptr<iface::engine>
@@ -45,16 +46,19 @@ namespace metashell
           logger* logger_) const;
 
     const std::string& args() const;
-    const std::string& description() const;
+    const data::markdown_string& description() const;
 
     const std::vector<data::feature>& features() const;
 
   private:
     engine_factory _factory;
     std::string _args;
-    std::string _description;
+    data::markdown_string _description;
     std::vector<data::feature> _features;
   };
+
+  std::string list_features(const engine_entry& engine_);
+  data::markdown_string list_features_in_markdown(const engine_entry& engine_);
 }
 
 #endif